coordlist is int

+++ com/BombingGames/WurfelEngine/Core/Map/AbstractPosition.java
@@ -37,8 +37,8 @@ import com.badlogic.gdx.math.Vector3;
  * @author Benedikt Vogler
  */
 public abstract class AbstractPosition {
-    private final byte referenceX;//top left chunk x coordinate
-    private final byte referenceY;//top left chunk Y coordinate
+    private final int referenceX;//top left chunk x coordinate
+    private final int referenceY;//top left chunk Y coordinate
     private float height;
 
     /**
@@ -46,7 +46,7 @@ public abstract class AbstractPosition {
      * @param topleftX the chunk's X coordinate of the chunk at the top left
      * @param topleftY the chunk's Y coordinate of the chunk at the top left 
      */
-    public AbstractPosition(byte topleftX, byte topleftY) {
+    public AbstractPosition(int topleftX, int topleftY) {
         this.referenceX = topleftX;
         this.referenceY = topleftY;
     }
@@ -80,7 +80,7 @@ public abstract class AbstractPosition {
     * 
     * @return 
     */
-    protected byte getReferenceX() {
+    protected int getReferenceX() {
         return referenceX;
     }
 
@@ -88,7 +88,7 @@ public abstract class AbstractPosition {
      * 
      * @return 
      */
-    protected byte getReferenceY() {
+    protected int getReferenceY() {
         return referenceY;
     }
